requirements.txt
argh==0.26.2
pathtools==0.1.2
PyYAML==5.3
requests==2.22.0
watchdog==0.9.0
